requirements.txt
# Core
openai>=1.12.0
chromadb>=0.4.22
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# Data fetching
requests>=2.31.0

# UI
streamlit>=1.32.0

# Environment
python-dotenv>=1.0.0
